requirements.txt
django
djangorestframework
psycopg2
django-filter
markdown
drf-yasg